requirements.txt
channels==4.3.1
cryptography==45.0.6
daphne==4.2.1
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.11.0
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
openai==1.102.0
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
# Azure Blob Storage
azure-storage-blob==12.26.0